deployment scripts for adage v0.0.1
No more steps.sh: we are all in on fabric deployment now!
Get a PostgreSQL database running on an AWS RDS instance, set up
your config.py (in adage-server) to point there, fill in the other
AWS configuration parameters and options you want in config.py and
then type "fab" to invoke this file. An EC2 instance will be created
(using the AWS IAM credentials you supplied in config.py) and
This script will take over, running all deployment steps on its own.

Two points of user intervention are required, so this is not quite
a fire and forget scenario:

1. the script will tell the server to generate a key pair for code
deployment. The public key from this pair needs to be uploaded to
bitbucket so the server can pull code from our private repository.

2. after the keys are uploaded and the script is resumed, the server
will prompt you to add the bitbucket key to your known_hosts file. Type
yes and then the rest of the deployment should continue on its own.

If everything succeeded, you'll need to (re)assign the AWS elastic IP
address to the new instance.

Lots and lots of changes in the details for each deployment task,
some new tasks, and we now depend upon the fabric scripts in the
adage-server project so that we can avoid duplicating code.